backend/pyproject.toml
[project]
name = "survivor-pool-api"
version = "0.1.0"
description = "API for the Survivor Pool App"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "pymongo>=4.14.1",
    "bcrypt>=4.2.1",
    "pydantic[email]>=2.10.3",
    "sendgrid>=6.11.0",
    "pyjwt>=2.10.0",
]

[dependency-groups]
dev = [
    "ipython>=9.5.0",
    "ruff>=0.12.10",
]

[tool.uv]
python-preference = "only-system"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "N",   # pep8-naming (naming conventions)
    "S",   # flake8-bandit (security issues)
    "T20", # flake8-print (catches leftover print statements)
    "RUF", # Ruff-specific rules
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
]
